moved the signin, signup, forgetpassword to Auth directory
